refactored tests isolating those linux-only
