+++ openwhisk/initHandler_test.go
@@ -41,26 +41,6 @@ func Example_json_init() {
 	// {"value":{"code":"MQo=","binary":true,"main":"hello"}}
 }
 
-func Example_badinit_nocompiler() {
-	ts, cur, log := startTestServer("")
-	sys("_test/build.sh")
-	doRun(ts, "")
-	doInit(ts, "{}")
-	//sys("ls", "_test/exec")
-	doInit(ts, initBinary("_test/exec", ""))      // empty
-	doInit(ts, initBinary("_test/hi", ""))        // say hi
-	doInit(ts, initBinary("_test/hello.src", "")) // source not excutable
-	doRun(ts, "")
-	stopTestServer(ts, cur, log)
-	// Output:
-	// 400 {"error":"no action defined yet"}
-	// 200 {"ok":true}
-	// 400 {"error":"cannot start action: command exited"}
-	// 400 {"error":"cannot start action: command exited"}
-	// 400 {"error":"cannot start action: command exited"}
-	// 400 {"error":"no action defined yet"}
-}
-
 func Example_bininit_nocompiler() {
 	ts, cur, log := startTestServer("")
 	sys("_test/build.sh")
@@ -107,7 +87,7 @@ func Example_zipinit_nocompiler() {
 	// XXX_THE_END_OF_A_WHISK_ACTIVATION_XXX
 }
 
-/* commented out test for timing problems
+/* commented out test for timing problems*/
 func Example_shell_nocompiler() {
 	ts, cur, log := startTestServer("")
 	doRun(ts, "")
@@ -126,7 +106,7 @@ func Example_shell_nocompiler() {
 	// XXX_THE_END_OF_A_WHISK_ACTIVATION_XXX
 	// Goodbye!
 	// XXX_THE_END_OF_A_WHISK_ACTIVATION_XXX
-}*/
+} /**/
 
 func Example_main_nocompiler() {
 	ts, cur, log := startTestServer("")
